only send when configured to

+++ OpenAlprWebhookProcessor/WebPushSubscriptions/WebPushNotificationProducer.cs
@@ -35,12 +35,12 @@ namespace OpenAlprWebhookProcessor.WebPushSubscriptions
             _serviceProvider = serviceProvider;
             _logger = logger;
 
-            using(var scope = _serviceProvider.CreateScope())
+            using (var scope = _serviceProvider.CreateScope())
             {
                 var processorContext = scope.ServiceProvider.GetRequiredService<ProcessorContext>();
 
                 var keys = VapidKeyHelper.GetVapidKeys(processorContext);
-                
+
                 _pushClient.DefaultAuthentication = new VapidAuthentication(
                     keys.PublicKey,
                     keys.PrivateKey)
@@ -58,27 +58,38 @@ namespace OpenAlprWebhookProcessor.WebPushSubscriptions
             AlertUpdateRequest alert,
             CancellationToken cancellationToken)
         {
-            PushMessage notification = new AngularWebPushNotification
+            using (var scope = _serviceProvider.CreateScope())
             {
-                Title = $"Plate Seen: {alert.PlateNumber}",
-                Body = $"Plate {alert.PlateNumber} seen at {DateTimeOffset.UtcNow:g}",
-                Icon = "assets/icons/icon-96x96.png",
-                Image = alert.PlateJpegUrl,
-            }.ToPushMessage();
+                var processorContext = scope.ServiceProvider.GetRequiredService<ProcessorContext>();
+                var clientSettings = await processorContext.WebPushSettings
+                    .AsNoTracking()
+                    .FirstOrDefaultAsync(cancellationToken);
 
-            foreach (PushSubscription subscription in _pushSubscriptionsService.GetAll())
-            {
-                try
+                if (clientSettings.IsEnabled && (alert.IsUrgent || clientSettings.SendEveryPlateEnabled))
                 {
-                    await _pushClient.RequestPushMessageDeliveryAsync(
-                        subscription,
-                        notification,
-                        cancellationToken);
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Failed to send WebPush message");
-                    _pushSubscriptionsService.Delete(subscription.Endpoint);
+                    PushMessage notification = new AngularWebPushNotification
+                    {
+                        Title = $"Plate Seen: {alert.PlateNumber}",
+                        Body = $"Plate {alert.PlateNumber} seen at {DateTimeOffset.UtcNow:g}",
+                        Icon = "assets/icons/icon-96x96.png",
+                        Image = alert.PlateJpegUrl,
+                    }.ToPushMessage();
+
+                    foreach (PushSubscription subscription in _pushSubscriptionsService.GetAll())
+                    {
+                        try
+                        {
+                            await _pushClient.RequestPushMessageDeliveryAsync(
+                                subscription,
+                                notification,
+                                cancellationToken);
+                        }
+                        catch (Exception ex)
+                        {
+                            _logger.LogError(ex, "Failed to send WebPush message");
+                            _pushSubscriptionsService.Delete(subscription.Endpoint);
+                        }
+                    }
                 }
             }
         }
@@ -99,7 +110,7 @@ namespace OpenAlprWebhookProcessor.WebPushSubscriptions
 
                 if (!credentialsValid)
                 {
-                    _logger.LogError("WebPush credentials are missing.");   
+                    _logger.LogError("WebPush credentials are missing.");
                 }
                 else
                 {
